Create new usuario package for user classes
