+++ src/library/Emprestimo.java
@@ -1,4 +1,5 @@
 package library;
+import usuario.*;
 
 import java.time.LocalDate;
