--- src/library/Funcionario.java
@@ -1,18 +1,0 @@
-package library;
-
-import java.time.LocalDate;
-import java.util.ArrayList;
-import java.util.LinkedList;
-import java.util.List;
-
-public class Funcionario extends Membro{
-
-    public Funcionario(String nome, int id, String endereco, String contato, LocalDate dataRegistro,
-            ArrayList<Multa> multas, List<Emprestimo> emprestimos) {
-        super(nome, id, endereco, contato, dataRegistro, multas, emprestimos);
-        
-    }
-
-    
-    
-}
