+++ src/library/Main.java
@@ -1,6 +1,8 @@
 package library;
 import java.util.ArrayList;
 import java.util.HashMap;
+import usuario.Membro;
+
 
 
 public class Main {
