+++ src/usuario/Administrador.java
@@ -1,0 +1,19 @@
+package usuario;
+
+import java.time.LocalDate;
+import java.util.ArrayList;
+import java.util.LinkedList;
+import java.util.List;
+
+import library.Emprestimo;
+import library.Multa;
+
+public class Administrador extends Funcionario{
+
+    public Administrador(String nome, int id, String endereco, String contato, LocalDate dataRegistro,
+            ArrayList<Multa> multas, List<Emprestimo> emprestimos) {
+        super(nome, id, endereco, contato, dataRegistro, multas, emprestimos);
+        
+    }
+    
+}
